+++ ttcn3-dumpcap-stop.sh
@@ -5,18 +5,17 @@ PIDFILE_NETCAT=/tmp/netcat.pid
 TESTCASE=$1
 VERDICT="$2"
 
-kill_rm_pidfile() {
-if [ -e $1 ]; then
-        PSNAME="$(ps -q "$(cat "$1")" -o comm=)"
-	if [ "$PSNAME" != "sudo" ]; then
-		kill "$(cat "$1")"
-	else
-	# NOTE: This requires you to be root or something like
-	# "laforge ALL=NOPASSWD: /usr/sbin/tcpdump, /bin/kill" in your sudoers file
-		sudo kill "$(cat "$1")"
-	fi
-	rm $1
+if ! [ "$(id -u)" = "0" ]; then
+	SUDOSTR="sudo -n"
+else
+	SUDOSTR=""
 fi
+
+kill_rm_pidfile() {
+	if ! [ -e "$1" ] && [ -s "$1" ]; then
+		$SUDOSTR kill "$(cat "$1")" 2>&1 || grep -v "No such process"
+		rm $1
+	fi
 }
 
 date
